modified:   example-bare/App.tsx 	new file:   example-bare/examples/Search/Header.tsx 	new file:   example-bare/examples/Search/SearchList.tsx 	new file:   example-bare/examples/Search/SearchResult.tsx 	modified:   example-bare/examples/Search/index.tsx 	modified:   example-bare/examples/Search/styles.ts 	modified:   example-bare/examples/TableOfContents/TableOfContents.tsx 	modified:   example-bare/examples/TableOfContents/index.tsx 	modified:   example-bare/examples/index.tsx 	modified:   example-bare/package.json 	modified:   example-bare/yarn.lock 	modified:   example-expo/App.tsx 	modified:   example-expo/examples/Bookmarks/BookmarksList.tsx 	deleted:    example-expo/examples/Search/ChapterList.tsx 	deleted:    example-expo/examples/Search/SeachList.tsx 	new file:   example-expo/examples/Search/SearchList.tsx 	modified:   example-expo/examples/Search/index.tsx 	modified:   example-expo/examples/TableOfContents/TableOfContents.tsx 	modified:   example-expo/examples/TableOfContents/index.tsx 	modified:   example-expo/examples/index.tsx 	modified:   example-expo/package.json 	modified:   example-expo/yarn.lock 	modified:   package.json 	modified:   src/context.tsx 	modified:   src/types.ts
