+++ example-bare/examples/Search/styles.ts
@@ -6,22 +6,4 @@ export const styles = StyleSheet.create({
     justifyContent: 'space-around',
     alignItems: 'center',
   },
-  options: {
-    width: '100%',
-    flexDirection: 'column',
-    justifyContent: 'space-around',
-    alignItems: 'center',
-    padding: 8,
-  },
-  currentFormat: {
-    textAlign: 'center',
-  },
-  form: {
-    width: '100%',
-  },
-  input: {
-    backgroundColor: '#c0c0c0',
-    borderRadius: 15,
-    width: '100%',
-  },
 });
