+++ example-expo/examples/index.tsx
@@ -6,5 +6,5 @@ export * from './Search';
 export * from './OpenExternalLink';
 export * from './Annotations';
 export * from './Bookmarks';
-export * from './TableOfContents';
+export { default as TableOfContents } from './TableOfContents';
 export * from './JavascriptInjection';
